# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "Documentation for e-commerce APIMy API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Get All Brands",
        "produces": [
          "text/json"
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Tablets"
                              },
                              "createdAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              },
                              "updatedAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Add a Category",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tablet"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "result": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number",
                              "example": 7
                            },
                            "name": {
                              "type": "string",
                              "example": "test"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2023-11-30T07:37:23.428Z"
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2023-11-30T07:37:23.428Z"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Name is required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/categories/{categoryId}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Delete a category",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Category Id",
            "example": 1
          },
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "The Category is being used"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Edit a Category",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Category Id",
            "example": 1
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "test1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "string",
                          "example": "Category updated"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Name is required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}